+++ data_tests/saved__backend__/numba/add_inline.py
@@ -1,12 +1,12 @@
 # __protected__ from numba import njit
-# __protected__ @njit
+# __protected__ @njit(cache=True, fastmath=True)
 
 
 def add(a, b):
     return a + b
 
 
-# __protected__ @njit
+# __protected__ @njit(cache=True, fastmath=True)
 
 
 def use_add(n=10000):
@@ -16,4 +16,4 @@ def use_add(n=10000):
     return tmp
 
 
-__transonic__ = ("0.3.2",)
+__transonic__ = ("0.4.7",)