+++ data_tests/saved__backend__/numba/block_fluidsim.py
@@ -1,5 +1,5 @@
 # __protected__ from numba import njit
-# __protected__ @njit
+# __protected__ @njit(cache=True, fastmath=True)
 
 
 def rk2_step0(state_spect_n12, state_spect, tendencies_n, diss2, dt):
@@ -15,4 +15,4 @@ def rk2_step0(state_spect_n12, state_spect, tendencies_n, diss2, dt):
 arguments_blocks = {
     "rk2_step0": ["state_spect_n12", "state_spect", "tendencies_n", "diss2", "dt"]
 }
-__transonic__ = ("0.3.0.post0",)
+__transonic__ = ("0.4.7",)