+++ data_tests/saved__backend__/numba/blocks_type_hints.py
@@ -1,5 +1,5 @@
 # __protected__ from numba import njit
-# __protected__ @njit
+# __protected__ @njit(cache=True, fastmath=True)
 
 
 def block0(a, b, n):
@@ -16,4 +16,4 @@ def block0(a, b, n):
 
 
 arguments_blocks = {"block0": ["a", "b", "n"]}
-__transonic__ = ("0.4.2",)
+__transonic__ = ("0.4.7",)