+++ data_tests/saved__backend__/numba/boosted_class_use_import.py
@@ -2,7 +2,7 @@
 import numpy as np
 from exterior_import_boost import func_import
 
-# __protected__ @njit
+# __protected__ @njit(cache=True, fastmath=True)
 
 
 def __for_method__MyClass2__myfunc(self_attr0, self_attr1, arg):
@@ -10,4 +10,4 @@ def __for_method__MyClass2__myfunc(self_attr0, self_attr1, arg):
 
 
 __code_new_method__MyClass2__myfunc = "\n\ndef new_method(self, arg):\n    return backend_func(self.attr0, self.attr1, arg)\n\n"
-__transonic__ = ("0.3.0.post0",)
+__transonic__ = ("0.4.7",)