+++ data_tests/saved__backend__/numba/boosted_func_use_import.py
@@ -2,11 +2,11 @@
 import numpy as np
 from exterior_import_boost import func_import
 
-# __protected__ @njit
+# __protected__ @njit(cache=True, fastmath=True)
 
 
 def func(a, b):
     return (a * np.log(b)).max() + func_import()
 
 
-__transonic__ = ("0.3.0.post0",)
+__transonic__ = ("0.4.7",)