+++ data_tests/saved__backend__/numba/default_params.py
@@ -1,5 +1,5 @@
 # __protected__ from numba import njit
-# __protected__ @njit
+# __protected__ @njit(cache=True, fastmath=True)
 
 
 def func(a=1, b=None, c=1.0):
@@ -7,4 +7,4 @@ def func(a=1, b=None, c=1.0):
     return a + c
 
 
-__transonic__ = ("0.4.0",)
+__transonic__ = ("0.4.7",)