+++ data_tests/saved__backend__/numba/mixed_classic_type_hint.py
@@ -1,18 +1,18 @@
 # __protected__ from numba import njit
 import numpy as np
 
-# __protected__ @njit
+# __protected__ @njit(cache=True, fastmath=True)
 
 
 def func(a, b):
     return (a * np.log(b)).max()
 
 
-# __protected__ @njit
+# __protected__ @njit(cache=True, fastmath=True)
 
 
 def func1(a, b):
     return a * np.cos(b)
 
 
-__transonic__ = ("0.3.0.post0",)
+__transonic__ = ("0.4.7",)