+++ data_tests/saved__backend__/numba/no_arg.py
@@ -1,16 +1,16 @@
 # __protected__ from numba import njit
-# __protected__ @njit
+# __protected__ @njit(cache=True, fastmath=True)
 
 
 def func():
     return 1
 
 
-# __protected__ @njit
+# __protected__ @njit(cache=True, fastmath=True)
 
 
 def func2():
     return 1
 
 
-__transonic__ = ("0.3.0.post0",)
+__transonic__ = ("0.4.7",)